move preface files to components/Preface
